+++ .vscode-test.mjs
@@ -1,5 +1,9 @@
 import { defineConfig } from '@vscode/test-cli';
 
 export default defineConfig({
-	files: 'test/**/*.test.js',
+    files: 'out/test/**/*.test.js',
+    mocha: {
+        ui: 'tdd',
+        color: true
+    }
 });