+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 All notable changes to the "baseline-vscode" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.0.1] - 2026-03-22
 
-## [Unreleased]
+### Added
 
-- Initial release
+- Search for web platform features and get their Baseline info
+- Autocomplete support for [baseline-todo](https://github.com/rviscomi/baseline-todo) syntax
+- Autocomplete support for [`<baseline-status featureId>`](https://github.com/web-platform-dx/baseline-status)
+- Hover support for baseline-todo syntax to display Baseline info at a glance
+- Validation of feature IDs within baseline-todo syntax